created Match mixins; deleted manager; added mixins to views
